# lib/api-zod/src/generated/types/employeeClockStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeClockStatusStatus } from './employeeClockStatusStatus';

export interface EmployeeClockStatus {
  userId: number;
  name: string;
  /** @nullable */
  department: string | null;
  status: EmployeeClockStatusStatus;
  /** @nullable */
  clockedInAt: Date | null;
}