# lib/api-zod/src/generated/types/employeeStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestBreakdown } from './requestBreakdown';

export interface EmployeeStats {
  weeklyHours: number;
  avgDailyHours: number;
  requestBreakdown: RequestBreakdown;
}