# lib/api-zod/src/generated/types/shiftDay.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShiftDay {
  date: Date;
  dayName: string;
  plannedStart: string;
  plannedEnd: string;
  isToday: boolean;
}