# lib/api-zod/src/generated/types/timeEntryInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimeEntryInputType } from './timeEntryInputType';

export interface TimeEntryInput {
  type: TimeEntryInputType;
}