# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * HR Management API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserLanguage } from './userLanguage';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  profilePictureUrl?: string | null;
  language: UserLanguage;
  /** @nullable */
  shiftPreferences?: string | null;
  createdAt: Date;
}